Put in most of the constraints.  Realised that I need to be really careful with the dimensions of the various np.arrays (especially, e.g. row or columns or 1D!), and am down to _build_weighting_matrices() in updating the docstrings to make sure they are consistently explicit about this.
Note that there is also a units thing that needs checking - Josh's code
assumes he is working in radius and not depth, and m/s not km/s.  I think
this is just so as not to convert the MINEOS model card back and forth,
but at the least, it means I have mixed m/s and km/s in here - if not worse!

Still nothing is actually even tried out, really, in the constraints
equations section.  But it should at least be ready to try out.  Still have
a few constraints left to add - Josh's H7, H8, H9