Starting to set up the tests.  Trying to check that the G matrix works, and it is not promising so far...!
Note that I expected issues for changing the layer thickness, but so far
I have only perturbed velocities in situ.  So no good ideas (other than
the obvious - there is something wrong with my partial derivative equations
or the coding up thereof!) on what might be messing this up.

However, test architecture has been set up, with one trivial test (checking
that the default values work on define_models.SetupModel classes) and the
test for G.  This are NOT yet any tests for the actual code - this is straight
in at the deep end, i.e. checking the output of G * dm = dc (where G = dc/dm)
against the output of MINEOS, i.e. the difference between c calculated from
m_new and c calculated from m_old using MINEOS.

First trials suggest that I wasn't super far off for shallow changes to the
velocity model - although the polarity is wrong for G * dm, i.e. increasing
the shallow velocities predicted c to decrease throughout (though the effect
was the strongest on short periods - so that's something!!).  However,
changing the velocity node in the middle of the model went totally, with
no correlation between the two estimates of dc.

The code now exists to check for these differences (and plot them up), but
probably the most productive thing is to start running tests on the
existing code. So....  In for the long haul!

Note: also found some minor errors in the util codes already that
wouldn't let the tests run as written.  So the changes to all the core codes
are to make the thing run at all, rather than to make it run accurately!