This wasn't actually coded correctly to deal with when MINEOS doesn't just work the first time around.  Unfortunately, the model that I was trying with has now stopped breaking MINEOS, so didn't add it to the tests - should download Zach's model that breaks things and see if it works, but I think (?) it should now!  But should definitely add that case to test_pipeline.py.
Also replaced a bunch of statements that were try: delete something;
to if os.path.exists(something): delete something.  The try statement
is dangerous because it also hides the errors if they aren't the errors
you are expecting!

And there were things that were calling in global variables, so they
work in the context of the pipeline, but not if you are calling the
function alone e.g. eigfile.