Commiting before overhauling the damping....
Ok - one of the issues I may be having is that the damping is
completely fucked up.  I have this layer-specific damping
scaling which is n_model_points long (all depth points divided
into different layers, then two at the end for the BL depths)
Given H is a (n_constraint_eqs, n_model_params) matrix and h is
a (n_model_params, ) vector in the eq. H * m = h, this means that
every row of H is being multiplied element-wise by the scaling,
and h is multiplied element-wise by the scaling.  THIS MAKES NO
SENSE.  This means that we currently have...
   sc0 * H00 * m0 + sc1 * H01 * m1 + sc2 * H02 * m2 = sc0 * h0
as equivalent to H * m = h (H00 * m0 + H01 * m1 + H02 * m2 = h0)

Want sc0 * H00 * m0 + sc0 * H01 * m1 + sc0 * H02 * m2 = sc0 * h0
etc.  So if want to keep the layer-specific scaling, need to
make sure that H is square ALWAYS - i.e. can just assume that the
first row of H is a constraint on the first velocity node, the
second row of H is a constraint on the second vel node etc.

Note: I am not sure how big of a difference this will make, because
I have mostly (recently!) been playing with constant scaling for
a given set of constraints, in which case the bug will actually lead
to the same result as the whole of H and h are just being multiplied
by a single constant value.